# perf(build): speed up copying of shared and client files