1.Use `fast-glob` stream mode instead of synchronous mode. Synchronous mode blocks the main thread and affects performance.
2.Use `Promise.all` to copy files in parallel, not sequentially.
3.This code has some performance issues, using `await Promise.all` in the `forEach` will still block the thread.
https://github.com/vuejs/vitepress/blob/86a3c24d2a0efee2df4a742c7d5c329a0b6e4b10/scripts/copyShared.js#L4-L9

After each five times of docs-build time test, the build time without optimization is:
 1. 11.84s
 2. 12.5s
 3. 12.94s
 4. 12.56s
 5. 11.76s

the build time after optimization is:
1. 10.25s
2. 11.81s
3. 11.73s
4. 10.04s
5. 10.36s

windows 10
node v16.14.0