@brc-dd 2023-05-27
This might actually lead to race conditions. Suppose `fg.stream` emits `data` event for some file late, while the control reaches `await Promise.all`. In that case those file will not present in `copyPromises`.