@zonemeen 2023-05-27
I don't think it will lead to a race condition, because `Promise.all(copyPromises)` waits for all the Promise in the `copyPromises` array to finish executing.