@brc-dd 2023-05-27
Yeah that part I agree, but what if a promise is not there in `copyPromises`. `fg.stream(...).on('data', ...)` is async in itself.