@brc-dd 2023-05-27
Yeah. See:

* at time = 0, we get a file from stream, we push it to `copyPromises`
* at time = 1, the control reaches that `await Promise.all(...)` statement
* at time = 2, we get another file from stream, we push it to `copyPromises`, but that won't be copied