@zonemeen 2023-05-27
Yeah, I guess I have to write it like this.

```js
stream.on('end', async () => { 
   await Promise.all(copyPromises);
})
```