@brc-dd 2023-05-27
Ah yeah that should work, but is there any need 😅 these scripts are only used while locally debugging/releasing stuff.